Add support for loadbalancer interface
The current way that the relations are set up for CK, with
kubeapi-load-balancer sitting in between the master and workers, is
unnecessarily convoluted, and leads to mistakes due to the relations not
being set up properly since they have to be done differently depending
on whether or not kubeapi-load-balancer is being used. This change makes
this charm support the `loadbalancer` interface and depends on the
master to forward the LB address via the `kube-control` relation. This
means that there's only one relation between the workers and master, the
endpoint address is always handed off in the same way, and the relations
are consistent regardless of whether the API LB charm is being used or
if another provider (such as a cloud integrator charm) is providing the
LB.

Part of [lp:1897818][]

[lp:1897818]: https://bugs.launchpad.net/charmed-kubernetes-testing/+bug/1897818